Changed the filer list to default to simple search.
Also allowed upper case characters to be used since some urls use uppercase as ids and we can't catch them otherwise.

If we want to lowercase them, than we should probably allow the uppercase to be saved
and then lowercase them and make the regex case insensitive when we send it to the client.

If someone wants to do something special with regex just prefix the regex with 'regex:' and it won't be touched.

Finally, changed the debug file to debug-in and added a debug-out to see what we send the client.